+++ src/model/items/Bow.java
@@ -26,7 +26,7 @@ public class Bow extends AbstractNonMagicalItem {
    * @param maxRange
    *     the maximum range of the bow
    */
-  public Bow(final String name, final int power, final int minRange, final int maxRange) {
+  public Bow(final String name, final double power, final int minRange, final int maxRange) {
     super(name, power, minRange, maxRange);
     this.minRange = Math.max(minRange, 2);
     this.maxRange = Math.max(maxRange, this.minRange);
